requirements.txt
pandas~=2.1.4
PyYAML~=6.0.1
psycopg2-binary~=2.9.9